Refaktor: TilbakekrevingHendelse extender Hendelse med full tilbakekreving-struktur
- Flyttet TilbakekrevingHendelse til hendelser-pakken (krav for sealed class)
- Extender Hendelse med utførtAv = Applikasjon.Tilbakekreving
- Nestet Tilbakekreving data class speiler Kafka JSON-struktur
- Fjernet privat TilbakekrevingDTO fra TilbakekrevingMottak
- Rettet test for manglende key til å verifisere IllegalArgumentException
